bridgekeeper/challenge/requirements.txt
flask
requests
firebase-admin
pycrypto